In order to improve AerospikeClient and AsyncClient testability interfaces have been extracted and implementation classes renamed to respectively DefaultAerospikeClient and DefaultAsyncClient. This change was motivated by heavy use of 'final' in implementation.
